+++ src/app/cart/cart.module.ts
@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
-import { CartListComponent } from './components/cart-list/cart-list.component';
-import { CartItemComponent } from './components/cart-item/cart-item.component';
 import { SharedModule } from '../shared/shared.module';
+import { CartItemComponent, CartListComponent } from './components';