+++ src/app/core/services/index.ts
@@ -1,0 +1,6 @@
+export * from './config-options.service';
+export * from './constant.service';
+export * from './gen-id.generator';
+export * from './generator.factory';
+export * from './generator';
+export * from './local-storage.service';